# protos/runtimebase/openapi/runtime.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "runtime.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/plugins/pallet": {
      "get": {
        "operationId": "RuntimeService_GetPalletTree",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/RuntimePalletTree"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "RuntimeService"
        ]
      }
    },
...
    "/api/runtime": {
      "get": {
        "operationId": "RuntimeService_GetObjects",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/RuntimeObjectsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "withData",
            "in": "query",
            "required": false,
            "type": "boolean"
          }
        ],
        "tags": [
          "RuntimeService"
        ]
      },
      "post": {
        "operationId": "RuntimeService_ObjectsDeploy",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/RuntimeObjectDeploy"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RuntimeObjectDeploy"
            }
          }
        ],
        "tags": [
          "RuntimeService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "RuntimeConnection": {
      "type": "object",
      "properties": {
        "connectionUUID": {
          "type": "string"
        },
        "targetConnectionUUID": {
          "type": "string"
        },
        "sourceUUID": {
          "type": "string"
        },
        "sourcePort": {
          "type": "string"
        },
        "sourcePortUUID": {
          "type": "string"
        },
        "targetUUID": {
          "type": "string"
        },
        "targetPort": {
          "type": "string"
        },
        "targetPortUUID": {
          "type": "string"
        },
        "isExistingConnection": {
          "type": "boolean"
        },
        "flowDirection": {
          "type": "string"
        },
        "disable": {
          "type": "boolean"
        },
        "isError": {
          "type": "boolean"
        },
        "created": {
          "type": "string"
        },
        "lastOk": {
          "type": "string"
        },
        "lastFail": {
          "type": "string"
        },
        "failCount": {
          "type": "integer",
          "format": "int32"
        },
        "error": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "wiresheetUUID": {
          "type": "string"
        }
      }
    },
...
    "RuntimeInfo": {
      "type": "object",
      "properties": {
        "objectID": {
          "type": "string"
        },
        "objectType": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "pluginName": {
          "type": "string"
        },
        "workingGroup": {
          "type": "string"
        },
        "workingGroupLeader": {
          "type": "string"
        },
        "workingGroupParent": {
          "type": "string"
        },
        "icon": {
          "type": "string"
        },
        "workingGroupObjects": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "workingGroupChildObjects": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "metaTags": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "flags": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "permissions": {
          "$ref": "#/definitions/RuntimePermissions"
        },
        "requirements": {
          "$ref": "#/definitions/RuntimeRequirements"
        },
        "isVirtualObject": {
          "type": "boolean"
        },
        "help": {
          "type": "string"
        },
        "supportDynamicInputs": {
          "type": "boolean"
        },
        "dynamicInputsMaxLimit": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "RuntimeMeta": {
      "type": "object",
      "properties": {
        "objectUUID": {
          "type": "string"
        },
        "objectName": {
          "type": "string"
        },
        "parentUUID": {
          "type": "string"
        },
        "position": {
          "$ref": "#/definitions/RuntimePosition"
        },
        "dynamicInputsCount": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "RuntimeNestedObjectConfigMap": {
      "type": "object",
      "properties": {
        "entries": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/RuntimeObjectConfig"
          }
        }
      }
    },
    "RuntimeObjectConfig": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "info": {
          "$ref": "#/definitions/RuntimeInfo"
        },
        "inputs": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/RuntimePort"
          }
        },
        "outputs": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/RuntimePort"
          }
        },
        "meta": {
          "$ref": "#/definitions/RuntimeMeta"
        },
        "stats": {
          "$ref": "#/definitions/RuntimeObjectStats"
        },
        "connections": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/RuntimeConnection"
          }
        },
        "settings": {
          "$ref": "#/definitions/RuntimeObjectSettings"
        },
        "portValues": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/RuntimePortValue"
          }
        }
      }
    },
    "RuntimeObjectDeploy": {
      "type": "object",
      "properties": {
        "deleted": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "new": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/RuntimeObjectConfig"
          }
        },
        "updated": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/RuntimeObjectConfig"
          }
        }
      }
    },
...
    "RuntimeObjectSettings": {
      "type": "object",
      "properties": {
        "uuid": {
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "persist": {
          "type": "boolean"
        }
      }
    },
    "RuntimeObjectStats": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "loaded": {
          "type": "string"
        },
        "loopCount": {
          "type": "integer",
          "format": "int64"
        },
        "timeSince": {
          "type": "string"
        }
      }
    },
    "RuntimeObjectsResponse": {
      "type": "object",
      "properties": {
        "objects": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/RuntimeObjectConfig"
          }
        }
      }
    },
...
    "RuntimePalletTree": {
      "type": "object",
      "properties": {
        "plugins": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/RuntimePlugins"
          }
        }
      }
    },
...
    "RuntimePermissions": {
      "type": "object",
      "properties": {
        "allPermissions": {
          "type": "boolean"
        },
        "canBeCreated": {
          "type": "boolean"
        },
        "canBeDeleted": {
          "type": "boolean"
        },
        "canBeUpdated": {
          "type": "boolean"
        },
        "readOnly": {
          "type": "boolean"
        },
        "allowHotFix": {
          "type": "boolean"
        },
        "forceDelete": {
          "type": "boolean"
        }
      }
    },
...
    "RuntimePlugins": {
      "type": "object",
      "properties": {
        "drivers": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/RuntimeNestedObjectConfigMap"
          }
        },
        "rubixNetwork": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/RuntimeNestedObjectConfigMap"
          }
        },
        "logic": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/RuntimeNestedObjectConfigMap"
          }
        },
        "services": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/RuntimeNestedObjectConfigMap"
          }
        }
      }
    },
    "RuntimePort": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "portUUID": {
          "type": "string"
        },
        "direction": {
          "type": "string"
        },
        "dataType": {
          "type": "string"
        },
        "defaultPosition": {
          "type": "integer",
          "format": "int32"
        },
        "transformation": {
          "type": "object"
        },
        "enablePersistence": {
          "type": "boolean"
        },
        "PersistenceCount": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "RuntimePortValue": {
      "type": "object",
      "properties": {
        "objectUUID": {
          "type": "string"
        },
        "portID": {
          "type": "string"
        },
        "dataType": {
          "type": "string"
        },
        "isNil": {
          "type": "boolean"
        },
        "transformationApplied": {
          "type": "boolean"
        },
        "data": {
          "type": "string",
          "format": "byte"
        },
        "floatValue": {
          "type": "number",
          "format": "double"
        },
        "stringValue": {
          "type": "string"
        },
        "boolValue": {
          "type": "boolean"
        },
        "intValue": {
          "type": "integer",
          "format": "int32"
        },
        "jsonValue": {
          "type": "string"
        },
        "displayValue": {
          "type": "string"
        },
        "portIDs": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "RuntimePosition": {
      "type": "object",
      "properties": {
        "positionY": {
          "type": "integer",
          "format": "int32"
        },
        "positionX": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "RuntimeRequirements": {
      "type": "object",
      "properties": {
        "callResetOnDeploy": {
          "type": "boolean"
        },
        "allowRuntimeAccess": {
          "type": "boolean"
        },
        "maxOne": {
          "type": "boolean"
        },
        "mustLiveInObjectType": {
          "type": "boolean"
        },
        "mustLiveParent": {
          "type": "boolean"
        },
        "requiresLogger": {
          "type": "boolean"
        },
        "supportsActions": {
          "type": "boolean"
        },
        "servicesRequirements": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "loggerOptions": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}